requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
backports.zstd==1.3.0
beautifulsoup4==4.14.3
bleach==6.3.0
blinker==1.9.0
Brotli==1.2.0
cachetools==7.0.5
catboost==1.2.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.0
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
fsspec==2026.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.46
graphviz==0.21
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperopt==0.2.7
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lightgbm==4.6.0
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mistune==3.2.0
ml_dtypes==0.5.4
mpmath==1.3.0
munkres==1.1.4
narwhals==2.17.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest_asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numba==0.64.0
numpy==2.4.2
onnx==1.20.1
onnx-ir==0.2.0
onnxruntime==1.24.3
onnxscript==0.6.2
openai==2.30.0
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pandas-stubs==3.0.0.260204
pandocfilters==1.5.1
parso==0.8.5
pillow==12.1.0
pip==26.0.1
platformdirs==4.5.1
plotly==6.6.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pydotplus==2.0.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
PySide6==6.10.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
pytz==2026.1.post1
pywin32==311
pywinpty==3.0.3
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
runtime==0.1.4
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
Send2Trash==2.1.0
setuptools==81.0.0
shap==0.51.0
shiboken6==6.10.2
six==1.17.0
slicer==0.0.8
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
stack_data==0.6.3
streamlit==1.55.0
streamlit-audiorecorder==0.0.6
sympy==1.14.0
tenacity==9.1.4
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
torch==2.10.0
torchaudio==2.10.0+cpu
torchinfo==1.8.0
torchvision==0.25.0+cpu
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
unicodedata2==17.0.0
uri-template==1.3.0
urllib3==2.6.3
watchdog==6.0.0
wcwidth==0.5.3
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.46.3
widgetsnbextension==4.0.15

xgboost==3.2.0
